Allow selecting any open database in unlock dialog
When showing the database unlock dialog from browser or auto-type and
there's more than one open database, replace the filename label with
a drop-down box of all available files. Switching the filename here will
change the active tab of the main window and unlock it.

This is the first part of an implementation for
https://github.com/keepassxreboot/keepassxc/issues/2322

TODO/Limitations:
  * The drop-down list should show only locked databases rather than all
    open databases.
  * The drop-down should dynamically respond if the set of open database
    tabs is changed (e.g. if the user closes a database in the main
    window while the unlock dialog is open)
  * No handling for the case where the active database is unlocked but
    has no matching credentials, and there's other locked databases open
    ("Case 2" in the github issue)