DatabaseOpenDialog: make it a Modal dialog, list only locked databases
If the main UI can be used while a database open dialog is open, it
creates a bunch of corner cases around how to act when databases are
opened/closed, locked/unlocked, or the active tab is changed.

It's possible to hook all of these events up to slots and have
DatabaseTabWidget and DatabaseOpenDialog update each other, but the
implementation gets pretty messy.

Alternatively, if DatabaseOpenDialog is Modal then its existence blocks
input to the main application and we can avoid all of those corner
cases.

Also only add locked database tabs to the multi-file list, rather than
all open tabs.